+++ models/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+* Initial workflow mapping support.
+
 ## [1.0.1] - 2021-08-24
 
 ### Fixed